refactor: modularize resource week view components and implement custom hook for view data management
