+++ src/features/resource-calendar/components/week-view/index.tsx
@@ -1,7 +1,7 @@
 import type React from 'react'
 import { useSmartCalendarContext } from '@/hooks/use-smart-calendar-context'
-import { ResourceWeekHorizontal } from './resource-week-horizontal'
-import { ResourceWeekVertical } from './resource-week-vertical'
+import { ResourceWeekHorizontal } from './horizontal/resource-week-horizontal'
+import { ResourceWeekVertical } from './vertical/resource-week-vertical'
 
 export const ResourceWeekView: React.FC = () => {
 	const { orientation } = useSmartCalendarContext()
