--- src/features/resource-calendar/components/week-view/resource-week-vertical.tsx
@@ -1,236 +1,0 @@
-import type React from 'react'
-import { useMemo } from 'react'
-import { AllDayCell } from '@/components/all-day-row/all-day-cell'
-import { AllDayRow } from '@/components/all-day-row/all-day-row'
-import { AnimatedSection } from '@/components/animations/animated-section'
-import { HourLabel } from '@/components/hour-label/hour-label'
-import { ResourceCell } from '@/components/resource-cell'
-import type { BusinessHours } from '@/components/types'
-import { VerticalGrid } from '@/components/vertical-grid/vertical-grid'
-import { getViewHours } from '@/features/calendar/utils/view-hours'
-import { useSmartCalendarContext } from '@/hooks/use-smart-calendar-context'
-import type { Dayjs } from '@/lib/configs/dayjs-config'
-import { cn } from '@/lib/utils'
-import { getWeekDays } from '@/lib/utils/date-utils'
-
-export const ResourceWeekVertical: React.FC = () => {
-	const {
-		currentDate,
-		getVisibleResources,
-		firstDayOfWeek,
-		t,
-		businessHours,
-		hideNonBusinessHours,
-		hiddenDays,
-		weekViewGranularity,
-	} = useSmartCalendarContext()
-
-	const resources = getVisibleResources()
-	// Generate week days
-	const weekDays = useMemo(
-		() => getWeekDays(currentDate, firstDayOfWeek),
-		[currentDate, firstDayOfWeek]
-	)
-
-	const visibleDays = useMemo(
-		() =>
-			hiddenDays
-				? weekDays.filter((day) => !hiddenDays.has(day.day()))
-				: weekDays,
-		[weekDays, hiddenDays]
-	)
-
-	const hours = useMemo(
-		() =>
-			getViewHours({
-				referenceDate: currentDate,
-				businessHours,
-				hideNonBusinessHours,
-				allDates: weekDays,
-				resourceBusinessHours: resources
-					.map((r) => r.businessHours)
-					.filter(Boolean) as (BusinessHours | BusinessHours[])[],
-			}),
-		[currentDate, businessHours, hideNonBusinessHours, weekDays, resources]
-	)
-
-	const firstCol = useMemo(
-		() => ({
-			id: weekViewGranularity === 'hourly' ? 'time-col' : 'date-col',
-			days: weekViewGranularity === 'hourly' ? hours : weekDays,
-			day: undefined,
-			className:
-				'shrink-0 w-16 min-w-16 max-w-16 sticky left-0 bg-background z-20',
-			gridType:
-				weekViewGranularity === 'hourly' ? ('hour' as const) : ('day' as const),
-			noEvents: true,
-			renderCell: (date: Dayjs) => (
-				<div className="text-muted-foreground p-2 text-right text-[10px] sm:text-xs flex flex-col items-center">
-					{weekViewGranularity === 'hourly' ? (
-						<HourLabel date={date} />
-					) : (
-						<div className="text-muted-foreground p-2 text-right text-[10px] sm:text-xs flex flex-col items-center">
-							<span>{date.format('ddd')}</span>
-							<span>{date.format('M/D')}</span>
-						</div>
-					)}
-					<HourLabel date={date} />
-				</div>
-			),
-			cellHeight: weekViewGranularity === 'hourly' ? 60 : 72,
-		}),
-		[hours, weekViewGranularity, weekDays]
-	)
-
-	const columns = useMemo(
-		() =>
-			weekViewGranularity === 'hourly'
-				? resources.flatMap((resource) =>
-						visibleDays.map((day) => ({
-							id: `day-col-${day.format('YYYY-MM-DD')}-resource-${resource.id}`,
-							resourceId: resource.id,
-							resource,
-							day,
-							days: getViewHours({
-								referenceDate: day,
-								businessHours,
-								hideNonBusinessHours,
-								allDates: weekDays,
-								resourceBusinessHours: resources
-									.map((r) => r.businessHours)
-									.filter(Boolean) as (BusinessHours | BusinessHours[])[],
-							}),
-							gridType: 'hour' as const,
-							cellHeight: 60,
-						}))
-					)
-				: resources.map((resource) => ({
-						id: `week-col-resource-${resource.id}`,
-						day: undefined,
-						resourceId: resource.id,
-						days: weekDays,
-						gridType: 'day' as const,
-						cellHeight: 72,
-					})),
-		[
-			resources,
-			weekDays,
-			businessHours,
-			hideNonBusinessHours,
-			visibleDays,
-			weekViewGranularity,
-		]
-	)
-
-	return (
-		<VerticalGrid
-			allDayRow={
-				weekViewGranularity === 'hourly' ? (
-					<div className="flex">
-						<AllDayCell />
-						{resources.map((resource) => (
-							<AllDayRow
-								classes={{ cell: 'min-w-50' }}
-								days={visibleDays}
-								key={`resource-week-allday-row-${resource.id}`}
-								resource={resource}
-								showSpacer={false}
-							/>
-						))}
-					</div>
-				) : undefined
-			}
-			classes={{ header: weekViewGranularity === 'hourly' ? 'h-24' : 'h-12' }}
-			columns={[firstCol, ...columns]}
-			data-testid="resource-week"
-			gridType={weekViewGranularity === 'hourly' ? 'hour' : 'day'}
-		>
-			<div className="flex-1 flex flex-col">
-				{/* Resource header row */}
-				<div className="flex h-12">
-					<div className="shrink-0 w-16 border-r z-20 bg-background sticky left-0">
-						<span
-							className={cn(
-								'px-2 h-full w-full flex justify-center text-xs text-muted-foreground',
-								weekViewGranularity === 'hourly'
-									? 'items-end'
-									: 'items-center border-b'
-							)}
-						>
-							{t('week')}{' '}
-							{weekViewGranularity === 'daily' ? currentDate.week() : ''}
-						</span>
-					</div>
-
-					{resources.map((resource, index) => {
-						const key = `resource-week-header-${resource.id}-day`
-
-						return (
-							<AnimatedSection
-								className={cn(
-									'shrink-0 border-r last:border-r-0 border-b flex items-center text-center font-medium'
-								)}
-								delay={index * 0.05}
-								key={`${key}-animated`}
-								style={{
-									width:
-										weekViewGranularity === 'hourly'
-											? `calc(${visibleDays.length} * var(--spacing) * 50)`
-											: 'calc(var(--spacing) * 50)',
-								}}
-								transitionKey={`${key}-motion`}
-							>
-								<ResourceCell
-									className="h-full w-full flex-1"
-									resource={resource}
-								>
-									<div
-										className={cn(
-											'sticky text-sm font-medium truncate',
-											weekViewGranularity === 'hourly' ? 'left-1/2' : 'left-1'
-										)}
-									>
-										{resource.title}
-									</div>
-								</ResourceCell>
-							</AnimatedSection>
-						)
-					})}
-				</div>
-
-				{/* Date header row */}
-				{weekViewGranularity === 'hourly' && (
-					<div className="flex h-12">
-						<div className="shrink-0 w-16 border-r border-b z-20 bg-background sticky left-0">
-							<span className="px-2 h-full w-full flex justify-center items-start font-medium">
-								{currentDate.week()}
-							</span>
-						</div>
-						{columns.map((col, index) => {
-							const day = col.day
-							if (!day) return null
-							const key = `resource-week-header-${day.toISOString()}-hour-${col.resourceId}`
-
-							return (
-								<AnimatedSection
-									className={cn(
-										'w-50 border-r last:border-r-0 border-b flex flex-col items-center justify-center text-xs shrink-0 bg-background'
-									)}
-									data-testid={`resource-week-time-label-${day.format('HH')}`}
-									delay={index * 0.05}
-									key={`${key}-animated`}
-									transitionKey={`${key}-motion`}
-								>
-									<div className="text-sm">{day.format('ddd')}</div>
-									<div className="text-xs text-muted-foreground">
-										{day.format('M/D')}
-									</div>
-								</AnimatedSection>
-							)
-						})}
-					</div>
-				)}
-			</div>
-		</VerticalGrid>
-	)
-}
